Separate helper classes out into separate file
Move the classes used in the migrate_with_timeouts command out into a
separate class, with the intention of enabling this to be used in a more
general way.
